[FIX] Resolve issue sonar
