--- src/main/java/id/ac/ui/cs/advprog/everest/modules/technicianReport/exception/DatabaseException.java
@@ -1,11 +1,0 @@
-package id.ac.ui.cs.advprog.everest.modules.technicianReport.exception;
-
-public class DatabaseException extends RuntimeException {
-    public DatabaseException(String message) {
-        super(message);
-    }
-
-    public DatabaseException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
