--- src/main/java/id/ac/ui/cs/advprog/everest/modules/technicianReport/exception/IllegalStateTransitionException.java
@@ -1,7 +1,0 @@
-package id.ac.ui.cs.advprog.everest.modules.technicianReport.exception;
-
-public class IllegalStateTransitionException extends RuntimeException {
-    public IllegalStateTransitionException(String currentState, String attemptedAction) {
-        super("Cannot perform " + attemptedAction + " action while in " + currentState + " state");
-    }
-}
