--- src/main/java/id/ac/ui/cs/advprog/everest/modules/technicianReport/exception/InvalidTechnicianReportStateException.java
@@ -1,12 +1,0 @@
-package id.ac.ui.cs.advprog.everest.modules.technicianReport.exception;
-
-public class InvalidTechnicianReportStateException extends RuntimeException {
-
-    public InvalidTechnicianReportStateException(String message) {
-        super(message);
-    }
-
-    public InvalidTechnicianReportStateException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
