--- src/main/java/id/ac/ui/cs/advprog/everest/modules/technicianReport/model/state/CompletedState.java
@@ -1,10 +1,0 @@
-package id.ac.ui.cs.advprog.everest.modules.technicianReport.model.state;
-
-import id.ac.ui.cs.advprog.everest.modules.technicianReport.model.TechnicianReport;
-
-public class CompletedState extends AbstractReportState {
-    @Override
-    public String getName() {
-        return "COMPLETED";
-    }
-}
