--- src/main/java/id/ac/ui/cs/advprog/everest/modules/technicianReport/model/state/RejectedState.java
@@ -1,8 +1,0 @@
-package id.ac.ui.cs.advprog.everest.modules.technicianReport.model.state;
-
-public class RejectedState extends AbstractReportState {
-    @Override
-    public String getName() {
-        return "REJECTED";
-    }
-}
