--- src/main/java/id/ac/ui/cs/advprog/everest/modules/technicianReport/model/state/SubmittedState.java
@@ -1,20 +1,0 @@
-package id.ac.ui.cs.advprog.everest.modules.technicianReport.model.state;
-
-import id.ac.ui.cs.advprog.everest.modules.technicianReport.model.TechnicianReport;
-
-public class SubmittedState extends AbstractReportState {
-    @Override
-    public String getName() {
-        return "SUBMITTED";
-    }
-
-    @Override
-    public ReportState approve(TechnicianReport context) {
-        return new ApprovedState();
-    }
-
-    @Override
-    public ReportState reject(TechnicianReport context) {
-        return new RejectedState();
-    }
-}
