+++ src/main/java/id/ac/ui/cs/advprog/everest/modules/technicianreport/exception/DatabaseException.java
@@ -1,0 +1,11 @@
+package id.ac.ui.cs.advprog.everest.modules.technicianreport.exception;
+
+public class DatabaseException extends RuntimeException {
+    public DatabaseException(String message) {
+        super(message);
+    }
+
+    public DatabaseException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
