+++ src/main/java/id/ac/ui/cs/advprog/everest/modules/technicianreport/exception/IllegalAccessTechnicianReport.java
@@ -1,0 +1,7 @@
+package id.ac.ui.cs.advprog.everest.modules.technicianreport.exception;
+
+public class IllegalAccessTechnicianReport extends RuntimeException {
+    public IllegalAccessTechnicianReport(String role, String action) {
+        super(role + " not authorized" + action);
+    }
+}
