+++ src/main/java/id/ac/ui/cs/advprog/everest/modules/technicianreport/exception/InvalidDataTechnicianReport.java
@@ -1,0 +1,7 @@
+package id.ac.ui.cs.advprog.everest.modules.technicianreport.exception;
+
+public class InvalidDataTechnicianReport extends RuntimeException {
+    public InvalidDataTechnicianReport(String message) {
+        super("Technician Report Data Error: " + message);
+    }
+}
