+++ src/main/java/id/ac/ui/cs/advprog/everest/modules/technicianreport/model/state/AbstractReportState.java
@@ -1,0 +1,49 @@
+package id.ac.ui.cs.advprog.everest.modules.technicianreport.model.state;
+
+import id.ac.ui.cs.advprog.everest.modules.technicianreport.exception.IllegalAccessTechnicianReport;
+import id.ac.ui.cs.advprog.everest.modules.technicianreport.exception.IllegalStateTransitionException;
+import id.ac.ui.cs.advprog.everest.modules.technicianreport.model.TechnicianReport;
+
+public abstract class AbstractReportState implements ReportState {
+    @Override
+    public ReportState submit(TechnicianReport context) {
+        throw new IllegalStateTransitionException(getName(), "submit");
+    }
+
+    @Override
+    public ReportState approve(TechnicianReport context) {
+        throw new IllegalStateTransitionException(getName(), "approve");
+    }
+
+    @Override
+    public ReportState reject(TechnicianReport context) {
+        throw new IllegalStateTransitionException(getName(), "reject");
+    }
+
+    @Override
+    public ReportState startWork(TechnicianReport context) {
+        throw new IllegalStateTransitionException(getName(), "startWork");
+    }
+
+    @Override
+    public ReportState complete(TechnicianReport context) {
+        throw new IllegalStateTransitionException(getName(), "complete");
+    }
+
+    @Override
+    public boolean technicianCanModify() {
+        return false;
+    }
+
+    @Override
+    public boolean customerCanSee() {
+        return true;
+    }
+
+    @Override
+    public void readPermissions(TechnicianReport context) {
+        if (!customerCanSee()) {
+            throw new IllegalAccessTechnicianReport("Customer", "see report in draft state");
+        }
+    }
+}
