+++ src/main/java/id/ac/ui/cs/advprog/everest/modules/technicianreport/model/state/CompletedState.java
@@ -1,0 +1,8 @@
+package id.ac.ui.cs.advprog.everest.modules.technicianreport.model.state;
+
+public class CompletedState extends AbstractReportState {
+    @Override
+    public String getName() {
+        return "COMPLETED";
+    }
+}
