+++ src/main/java/id/ac/ui/cs/advprog/everest/modules/technicianreport/model/state/RejectedState.java
@@ -1,0 +1,8 @@
+package id.ac.ui.cs.advprog.everest.modules.technicianreport.model.state;
+
+public class RejectedState extends AbstractReportState {
+    @Override
+    public String getName() {
+        return "REJECTED";
+    }
+}
