+++ src/main/java/id/ac/ui/cs/advprog/everest/modules/technicianreport/service/TechnicianReportAuditLogger.java
@@ -1,0 +1,12 @@
+package id.ac.ui.cs.advprog.everest.modules.technicianreport.service;
+
+import org.springframework.scheduling.annotation.Async;
+import org.springframework.stereotype.Component;
+
+@Component
+public class TechnicianReportAuditLogger {
+    @Async
+    public void logReportAction(String action, String reportId, String technicianId) {
+        System.out.printf("Audit: %s on report %s by technician %s%n", action, reportId, technicianId);
+    }
+}
